TCPCL: Move Client State code to own files
